Allow removing added file
