+++ src/translations/nbTranslations.json
@@ -333,10 +333,6 @@
   "default_point_of_contact": "<p><link1>Ta kontakt med Sikt</link1> hvis ingen er kontaktpunkt for resultatet.</p>",
   "details": "Detaljer",
   "disciplines": {
-    "0001": "Humaniora",
-    "0002": "Samfunnsvitenskap",
-    "0003": "Medisin og helsefag",
-    "0004": "Realfag og teknologi",
     "1003": "Arkeologi og konservering",
     "1005": "Asiatiske og afrikanske studier",
     "1007": "Engelsk",
@@ -408,7 +404,11 @@
     "1162": "Konstruksjonsfag",
     "1166": "Filosofi",
     "1168": "Historie og idéhistorie",
-    "1170": "Kunst, design og arkitektur"
+    "1170": "Kunst, design og arkitektur",
+    "0001": "Humaniora",
+    "0002": "Samfunnsvitenskap",
+    "0003": "Medisin og helsefag",
+    "0004": "Realfag og teknologi"
   },
   "does_not_match": "Matcher ikke",
   "editor": {
@@ -2282,5 +2282,6 @@
   },
   "unpublished_result": "Avpublisert resultat",
   "view_contact_info": "Se kontaktinformasjon",
-  "vocabulary_missing": "Institusjonen har ikke lagt til vokabular."
+  "vocabulary_missing": "Institusjonen har ikke lagt til vokabular.",
+  "add_file": "Legg til fil"
 }
